+++ pennylane_lightning/core/src/gates/GateOperation.hpp
@@ -57,6 +57,7 @@ enum class GateOperation : uint32_t {
     SingleExcitation,
     SingleExcitationMinus,
     SingleExcitationPlus,
+    PSWAP,
     /* Three-qubit gates */
     Toffoli,
     CSWAP,
@@ -98,6 +99,7 @@ enum class ControlledGateOperation : uint32_t {
     DoubleExcitation,
     DoubleExcitationMinus,
     DoubleExcitationPlus,
+    PSWAP,
     /* Multi-qubit gates */
     MultiRZ,
     GlobalPhase,